implement initialize pool
