+++ app/apptesting/concentrated_liquidity.go
@@ -1,6 +1,7 @@
 package apptesting
 
 import (
+	clmodel "github.com/osmosis-labs/osmosis/v13/x/concentrated-liquidity/model"
 	"github.com/osmosis-labs/osmosis/v13/x/concentrated-liquidity/types"
 )
 
@@ -12,7 +13,20 @@ var (
 
 // PrepareConcentratedPool sets up an eth usdc concentrated liquidity pool with pool ID 1, tick spacing of 1, and no liquidity
 func (s *KeeperTestHelper) PrepareConcentratedPool() types.ConcentratedPoolExtension {
-	pool, err := s.App.ConcentratedLiquidityKeeper.CreateNewConcentratedLiquidityPool(s.Ctx, 1, ETH, USDC, DefaultTickSpacing)
+	// Mint some assets to the account.
+	s.FundAcc(s.TestAccs[0], DefaultAcctFunds)
+
+	// Create a concentrated pool via the swaprouter
+	poolID, err := s.App.SwapRouterKeeper.CreatePool(s.Ctx, clmodel.NewMsgCreateConcentratedPool(s.TestAccs[0], ETH, USDC, DefaultTickSpacing))
 	s.Require().NoError(err)
+
+	// Retrieve the poolInterface via the poolID
+	poolI, err := s.App.ConcentratedLiquidityKeeper.GetPool(s.Ctx, poolID)
+	s.Require().NoError(err)
+
+	// Type cast the PoolInterface to a ConcentratedPoolExtension
+	pool, ok := poolI.(types.ConcentratedPoolExtension)
+	s.Require().True(ok)
+
 	return pool
 }
